+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/ZimManageViewModel.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package org.kiwix.kiwixmobile.zimManager
+package org.kiwix.kiwixmobile.zim_manager
 
 import android.app.Application
 import android.net.ConnectivityManager
@@ -42,7 +42,6 @@ import org.kiwix.kiwixmobile.core.data.remote.KiwixService
 import org.kiwix.kiwixmobile.core.downloader.model.DownloadModel
 import org.kiwix.kiwixmobile.core.entity.LibraryNetworkEntity
 import org.kiwix.kiwixmobile.core.entity.LibraryNetworkEntity.Book
-import org.kiwix.kiwixmobile.core.extensions.calculateSearchMatches
 import org.kiwix.kiwixmobile.core.extensions.registerReceiver
 import org.kiwix.kiwixmobile.core.utils.BookUtils
 import org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil
@@ -51,7 +50,7 @@ import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode.MULT
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode.NORMAL
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem.BookOnDisk
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState
 import org.kiwix.kiwixmobile.zim_manager.NetworkState.CONNECTED
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions.MultiModeFinished
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions.RequestDeleteMultiSelection
@@ -84,10 +83,10 @@ class ZimManageViewModel @Inject constructor(
   private val storageObserver: StorageObserver,
   private val kiwixService: KiwixService,
   private val context: Application,
-  private val connectivityBroadcastReceiver: org.kiwix.kiwixmobile.zimManager.ConnectivityBroadcastReceiver,
+  private val connectivityBroadcastReceiver: org.kiwix.kiwixmobile.zim_manager.ConnectivityBroadcastReceiver,
   private val bookUtils: BookUtils,
-  private val fat32Checker: org.kiwix.kiwixmobile.zimManager.Fat32Checker,
-  private val defaultLanguageProvider: org.kiwix.kiwixmobile.zimManager.DefaultLanguageProvider,
+  private val fat32Checker: org.kiwix.kiwixmobile.zim_manager.Fat32Checker,
+  private val defaultLanguageProvider: org.kiwix.kiwixmobile.zim_manager.DefaultLanguageProvider,
   private val dataSource: DataSource,
   private val connectivityManager: ConnectivityManager,
   private val sharedPreferenceUtil: SharedPreferenceUtil
