+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/fileselect_view/FileSelectListState.kt
@@ -15,17 +15,21 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
+package org.kiwix.kiwixmobile.zim_manager.fileselect_view
 
-package org.kiwix.kiwixmobile.zimManager
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode.NORMAL
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem.BookOnDisk
 
-import android.content.Context
-import org.kiwix.kiwixmobile.core.zim_manager.Language
-import javax.inject.Inject
-
-class DefaultLanguageProvider @Inject constructor(private val context: Context) {
-  fun provide() = Language(
-    context.resources.configuration.locale.isO3Language,
-    true,
-    1
-  )
+data class FileSelectListState(
+  val bookOnDiskListItems: List<BooksOnDiskListItem>,
+  val selectionMode: SelectionMode = NORMAL
+) {
+  val selectedBooks by lazy {
+    bookOnDiskListItems.filter(
+      BooksOnDiskListItem::isSelected
+    )
+      .filterIsInstance(BookOnDisk::class.java)
+  }
 }
