+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/fileselect_view/effects/NavigateToDownloads.kt
@@ -16,36 +16,15 @@
  *
  */
 
-package org.kiwix.kiwixmobile.zimManager.fileselect_view.effects
+package org.kiwix.kiwixmobile.zim_manager.fileselect_view.effects
 
-import android.net.Uri
-import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.FileProvider
-import androidx.core.os.bundleOf
-import org.kiwix.kiwixmobile.R
 import org.kiwix.kiwixmobile.core.base.SideEffect
 import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.navigate
-import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem.BookOnDisk
-import org.kiwix.kiwixmobile.localFileTransfer.URIS_KEY
+import org.kiwix.kiwixmobile.nav.destination.library.LocalLibraryFragmentDirections
 
-data class ShareFiles(private val selectedBooks: List<BookOnDisk>) :
-  SideEffect<Unit> {
+object NavigateToDownloads : SideEffect<Unit> {
   override fun invokeWith(activity: AppCompatActivity) {
-    val selectedFileContentURIs = selectedBooks.mapNotNull {
-      if (Build.VERSION.SDK_INT >= 24) {
-        FileProvider.getUriForFile(
-          activity,
-          activity.packageName + ".fileprovider",
-          it.file
-        )
-      } else {
-        Uri.fromFile(it.file)
-      }
-    }
-    activity.navigate(
-      R.id.localFileTransferFragment,
-      bundleOf(URIS_KEY to selectedFileContentURIs.toTypedArray())
-    )
+    activity.navigate(LocalLibraryFragmentDirections.actionNavigationLibraryToNavigationDownloads())
   }
 }
