+++ app/src/test/java/org/kiwix/kiwixmobile/zim_manager/library_view/adapter/LibraryListItemTest.kt
@@ -25,12 +25,11 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.kiwix.kiwixmobile.core.entity.LibraryNetworkEntity.Book
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState.CanWrite4GbFile
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState.CannotWrite4GbFile
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState.NotEnoughSpaceFor4GbFile
-import org.kiwix.kiwixmobile.zimManager.Fat32Checker.FileSystemState.Unknown
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState.CanWrite4GbFile
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState.CannotWrite4GbFile
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState.NotEnoughSpaceFor4GbFile
+import org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FileSystemState.Unknown
 import org.kiwix.kiwixmobile.zim_manager.library_view.adapter.LibraryListItem.BookItem
 
 internal class LibraryListItemTest {
@@ -51,7 +50,7 @@ internal class LibraryListItemTest {
 
   @Test
   internal fun `Unknown file system state greater than 4GB can't be downloaded`() {
-    every { book.getSize() } returns (org.kiwix.kiwixmobile.zimManager.Fat32Checker.FOUR_GIGABYTES_IN_KILOBYTES + 1).toString()
+    every { book.getSize() } returns (org.kiwix.kiwixmobile.zim_manager.Fat32Checker.FOUR_GIGABYTES_IN_KILOBYTES + 1).toString()
     assertThat(canBeDownloaded(book, Unknown)).isFalse()
   }
